[svn] Update progress code to use higher timer resolution. Message-ID: <m37k49oivp.fsf@hniksic.iskon.hr>

+++ src/ftp.c
@@ -69,7 +69,7 @@ typedef struct
   int st;			/* connection status */
   int cmd;			/* command code */
   struct rbuf rbuf;		/* control connection buffer */
-  long dltime;			/* time of the download */
+  double dltime;		/* time of the download in msecs */
   enum stype rs;		/* remote system reported by ftp server */ 
   char *id;			/* initial directory */
   char *target;			/* target file name */
